refactor: remove console log and move function generateToken to auth
